requirements.txt
fastapi
alembic
celery
unicorn
gunicorn
pydantic-settings
pydantic
psycopg2
aioredis
asyncpg
pre_commit
PyJWT
redis
requests
SQLAlchemy
websockets
sentry-sdk
aioboto3
firebase_admin
pytz
passlib
aio-pika